dynamic defaults
This introduces a new mechanism for dynamic defaults. If a struct is passed to
dflag, and there are already values present in the struct, we will use those
values instead of the 'default' value.

Precedence:

flag > struct value > struct tag

This enables workflows like this:

    type Foo struct {
        Bar string
    }

    flags := &Foo{"default"}
    // Foo.Bar will be "default" if not specified via flag
    dflag.Parse(flags)

(also -- thanks @samuelkarp for the suggestion here!)